+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
+    * FEATURE     #1517 [ContactBundle]   Added account DataProvider
     * FEATURE     #1513 [ContactBundle]   Added contact DataProvider
     * FEATURE     #1369 [ContentBundle]   Show icons in column navigation based on the user's permission
     * FEATURE     #1415 [ContentBundle]   Refactored SmartContent to use DataProvider to load content